requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
cloudinary==1.40.0
python-dotenv==1.0.1
gevent==24.2.1
gevent-websocket==0.10.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
